commit_shin_230102_4
BoardController_notice의 delete 유효성 체크

+++ src/main/java/com/wisetree/test/BoardController_notice.java
@@ -88,7 +88,12 @@ public class BoardController_notice {
 //			HttpServletRequest req,
 			@RequestParam(defaultValue = "0") int num) {
 		
-//		BoardVO vo=this.boardService.selectBoardByIdx(num);
+		BoardVO vo=this.boardService.selectBoardByIdx(num);
+		if(vo==null) {
+			return util.addMsgBack(m, "해당글은 존재하지 않아요");
+		}
+		
+		
 		int n=this.boardService.deleteBoard(num);
 		
 		String str=(n>0)?"글 삭제 성공":"삭제 실패";
